pyproject.toml
[project]
name = "pf-server"
version = "0.1.0"
description = "PF Server with OAuth authentication and per-user containers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docker>=7.1.0",
    "fastapi>=0.128.0",
    "fastmcp>=2.14.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "structlog>=24.0.0",
    "uvicorn[standard]==0.35",
    "websockets>=15.0.1",
    "backoff>=2.2.1",
    "rq>=2.6.1",
    "litellm==1.80",
    "rich>=13.0.0",
]

[project.scripts]
pf-cli = "cli.pf_cli:main"
pf-logs = "pf_server.log_viewer:main"

[project.optional-dependencies]
dev = [
    "asgi-lifespan>=2.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.8.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/pf_server", "src/pf_mcp"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
show_missing = true
fail_under = 0

[tool.ruff]
line-length = 88

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
